WIP: storage using XDR types
Seems I'm going to revert and not use XDR types. They add a whole bunch
of conversions that may confuse the verifier and need to be implemented.
On the other hand using XDR type might be better for debugging failed
verification attempts and expressing pre and post-conditions on ledger state.